% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analysis.R
\name{correlation_analysis}
\alias{correlation_analysis}
\title{correlation analysis - returns correlation matrix as a dataframe}
\usage{
correlation_analysis(
  geneset_file_path,
  x_indices,
  y_indices,
  corr_method = "pearson",
  corr_adj_method = "holm",
  create_corrplot = FALSE,
  write_results = FALSE,
  corrplot_file_name = "Corrplot.plot",
  output_file_name = "R.corr.csv",
  output_dir_path = "Results/Correlation_analysis/"
)
}
\arguments{
\item{geneset_file_path}{full file path of geneset to perform correlation on}

\item{x_indices}{subset of columns from geneset to be used as x value to obtain correlation}

\item{y_indices}{subset of columns from geneset to be used as y value to obtain correlation}

\item{corr_method}{correlation method}

\item{corr_adj_method}{p value adjustment method}

\item{create_corrplot}{boolean indicating if corrplot should be created}

\item{write_results}{boolean indicating if results should be written to file}

\item{corrplot_file_name}{corrplot file name}

\item{output_file_name}{results file name}

\item{output_dir_path}{file path for all outputs - results and plots}
}
\description{
correlation analysis - returns correlation matrix as a dataframe
}
