% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analysis.R
\name{create_connected_genes_plot}
\alias{create_connected_genes_plot}
\title{create connected genes plot}
\usage{
create_connected_genes_plot(
  corr_data,
  plot_file_name,
  output_dir_path = "Results/Correlation_analysis/"
)
}
\arguments{
\item{corr_data}{correlation data based on which to create network}

\item{plot_file_name}{plot file name}

\item{output_dir_path}{file path for all outputs - results and plots}
}
\description{
create connected genes plot
}
