% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{dRNASb}
\alias{dRNASb}
\title{dRNASb pipeline}
\usage{
dRNASb(
  data_file_path,
  phenotype_file_path,
  annotation_function_file_path,
  ppi_file_path,
  result_file_prefix = ""
)
}
\arguments{
\item{data_file_path}{full file path of input data file containing read counts in (gene x samples) format}

\item{phenotype_file_path}{full file path of phenotype file - contains mapping of sample to groups with groups column}

\item{annotation_function_file_path}{full file path of annotation function file}

\item{ppi_file_path}{full file path of ppi file}

\item{result_file_prefix}{a prefix string to be added to all results file}
}
\description{
dRNASb pipeline
}
